+++ src/OrderedList.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Bakame\Http\StructuredFields;
 
+use DateTimeInterface;
 use Iterator;
 use Stringable;
 use function array_filter;
@@ -31,7 +32,7 @@ final class OrderedList implements MemberList
      *
      * @return static
      */
-    public static function from(InnerList|Item|ByteSequence|Token|Stringable|bool|int|float|string ...$members): self
+    public static function from(InnerList|Item|ByteSequence|Token|DateTimeInterface|Stringable|bool|int|float|string ...$members): self
     {
         return self::fromList($members);
     }
@@ -49,7 +50,7 @@ final class OrderedList implements MemberList
         return $instance;
     }
 
-    private static function filterMember(StructuredField|ByteSequence|Token|Stringable|bool|int|float|string $member): InnerList|Item
+    private static function filterMember(StructuredField|ByteSequence|Token|DateTimeInterface|Stringable|bool|int|float|string $member): InnerList|Item
     {
         return match (true) {
             $member instanceof InnerList, $member instanceof Item => $member,
@@ -135,7 +136,7 @@ final class OrderedList implements MemberList
      *
      * @param InnerList<int, Item>|Item|DataType ...$members
      */
-    public function unshift(StructuredField|ByteSequence|Token|Stringable|bool|int|float|string ...$members): self
+    public function unshift(StructuredField|ByteSequence|Token|DateTimeInterface|Stringable|bool|int|float|string ...$members): self
     {
         $this->members = [...array_map(self::filterMember(...), array_values($members)), ...$this->members];
 
@@ -147,7 +148,7 @@ final class OrderedList implements MemberList
      *
      * @param InnerList<int, Item>|Item|DataType ...$members
      */
-    public function push(StructuredField|ByteSequence|Token|Stringable|bool|int|float|string ...$members): self
+    public function push(StructuredField|ByteSequence|Token|DateTimeInterface|Stringable|bool|int|float|string ...$members): self
     {
         $this->members = [...$this->members, ...array_map(self::filterMember(...), array_values($members))];
 
@@ -161,7 +162,7 @@ final class OrderedList implements MemberList
      *
      * @throws InvalidOffset If the index does not exist
      */
-    public function insert(int $index, StructuredField|ByteSequence|Token|Stringable|bool|int|float|string ...$members): self
+    public function insert(int $index, StructuredField|ByteSequence|Token|DateTimeInterface|Stringable|bool|int|float|string ...$members): self
     {
         $offset = $this->filterIndex($index);
         match (true) {
@@ -181,7 +182,7 @@ final class OrderedList implements MemberList
      *
      * @throws InvalidOffset If the index does not exist
      */
-    public function replace(int $index, StructuredField|ByteSequence|Token|Stringable|bool|int|float|string $member): self
+    public function replace(int $index, StructuredField|ByteSequence|Token|DateTimeInterface|Stringable|bool|int|float|string $member): self
     {
         if (null === ($offset = $this->filterIndex($index))) {
             throw InvalidOffset::dueToIndexNotFound($index);